backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
cryptography
requests
pydantic
pydantic-settings
python-multipart
